move v8 date parser to separate file
